Homework added... takes a while because of a slow site but works
couple things... twitter seems to be weird and sometimes doesnt work when scraping is done multiple times so please close your browser and start again if needed. sometimes it took a couple tries for twitter not to cause an error. also, the site the hemisphere images are scraped from is really slow for some reason to it takes a while for the scraping to finish. Thanks!